Fix typings
