+++ react/modules/inputValueType.ts
@@ -1,0 +1,5 @@
+export enum InputValueType {
+  'TEXT' = 'TEXT',
+  'BOOLEAN' = 'BOOLEAN',
+  'OPTIONS' = 'OPTIONS',
+}
